dependency_updater/requirements.txt
# Dependências da Cloud Function de Manutenção
functions-framework
google-cloud-firestore
google-cloud-secret-manager
PyGithub
requests
packaging